+++ causarray/__init__.py
@@ -7,13 +7,13 @@ if 'NUMBA_THREADING_LAYER' not in os.environ:
 os.environ.setdefault('KMP_WARNINGS', '0')
 
 __all__ = [
-    'LFC', #'ATE', 'SATE', 'FC', 
+    'LFC', 'gcate_lfc_batch', 'LFC_batch',
     'fit_glm', 'fit_glm_fast', 'fit_glm_ondisk',
-    'reset_random_seeds', 'fit_gcate'
+    'reset_random_seeds', 'fit_gcate', 'fit_gcate_batch',
     ]
 
 
-from causarray.DR_learner import LFC # ATE, SATE, FC
+from causarray.DR_learner import LFC, gcate_lfc_batch, LFC_batch  # ATE, SATE, FC
 from causarray.gcate_glm import fit_glm
 from causarray.nb_glm_fast import fit_glm_fast, fit_glm_ondisk
 from causarray.utils import prep_causarray_data, reset_random_seeds, comp_size_factor